chore(codex): address PR review requirements
