+++ src/locales/ar.json
@@ -592,6 +592,7 @@
         "runningSaveMessage": "نسخة Codex هذه قيد التشغيل حاليًا. سيؤدي استبدال الخدمة المحلية الآن إلى مقاطعة المحادثة الحالية.\n\nاختر «حفظ وإعادة تشغيل Codex» للتطبيق الآن، أو «حفظ لوقت لاحق» للاحتفاظ بالجلسة الحالية دون تغيير.",
         "saveAndRestart": "حفظ وإعادة تشغيل Codex",
         "saveForLater": "حفظ لوقت لاحق",
+        "savedForNextLaunch": "تم الحفظ. سيصبح ساريًا عند تشغيل Codex في المرة القادمة عبر Cockpit؛ وستبقى الجلسة الحالية دون تغيير.",
         "restartRollbackFailed": "فشل بدء المسار الجديد ولم نتمكن من إعادة تشغيل التكوين الأصلي: ",
         "restartRolledBack": "فشل بدء المسار الجديد. تم استعادة التكوين الأصلي وإعادة تشغيل Codex: {{error}}"
       }
