+++ src/locales/ru.json
@@ -545,6 +545,7 @@
         "runningSaveMessage": "Этот экземпляр Codex запущен. Замена локальной службы сейчас прервет текущий диалог.\n\nВыберите «Сохранить и перезапустить Codex», чтобы применить сейчас, или «Сохранить на потом», чтобы не прерывать текущий сеанс.",
         "saveAndRestart": "Сохранить и перезапустить Codex",
         "saveForLater": "Сохранить на потом",
+        "savedForNextLaunch": "Сохранено. Изменение вступит в силу при следующем запуске Codex через Cockpit; текущий сеанс останется без изменений.",
         "restartRollbackFailed": "Не удалось запустить новый маршрут, и исходную конфигурацию не удалось перезапустить: ",
         "restartRolledBack": "Не удалось запустить новый маршрут. Исходная конфигурация восстановлена, Codex перезапущен: {{error}}"
       }
